WindowsAppSDK-ProjectTemplates/python/unpackaged/pyproject.toml
[project]
name = "winapp-pick"
version = "0.1.0"
description = "Simple WinApp SDK file picker example"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "winappsdk-Foundation"
]

[project.scripts]
winapp-pick = "winapp_pick.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
find-links = ["../wheels"]